Role claim authorization Done

+++ Restaurant.Application/Dishess/Commands/DeleteDishes/DeleteDishesCommandHandler.cs
@@ -2,8 +2,10 @@
 using MediatR;
 using Microsoft.Extensions.Logging;
 using Restaurant.Application.Restaurantss.Commands.DeleteRestaurant;
+using Restaurant.Domain.Constants;
 using Restaurant.Domain.Entities;
 using Restaurant.Domain.Exceptions;
+using Restaurant.Domain.Interfaces;
 using Restaurant.Domain.Repositories;
 using System;
 using System.Collections.Generic;
@@ -15,7 +17,7 @@ namespace Restaurant.Application.Dishess.Commands.DeleteDishes
 {
     public class DeleteDishesCommandHandler(ILogger<DeleteDishesCommandHandler> logger,
          IRestaurantRepository restaurantsRepository,
-         IDishesRepository dishRepository, IMapper mapper)
+         IDishesRepository dishRepository, IMapper mapper, IRestaurantAuthorizationService restaurantAuthorizationService)
         : IRequestHandler<DeleteDishesCommand>
     {
         public async Task Handle(DeleteDishesCommand request, CancellationToken cancellationToken)
@@ -24,7 +26,10 @@ namespace Restaurant.Application.Dishess.Commands.DeleteDishes
             var restaurants = await restaurantsRepository.GetByIdAsync(request.RestaurantId);
             if (restaurants is null)
                 throw new NotFoundException(nameof(Restaurants), request.RestaurantId.ToString());
-           
+
+            if (!restaurantAuthorizationService.Authorize(restaurants, ResourceOperation.Update))
+                throw new ForbidException();
+
             await dishRepository.DeleteDishAsync(restaurants.Dishes);
         }
     }
